+++ src/domain/books/errors.ts
@@ -1,0 +1,6 @@
+export class BookSearchProviderError extends Error {
+  constructor(message: string) {
+    super(message);
+    this.name = "BookSearchProviderError";
+  }
+}